feat: auth + nav bar
